requirements.txt
# 在本地环境中，项目根目录下，执行命令收集所有包
# pip freeze > plist.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.3.0
diff-match-patch==20181111
Django==1.11.16
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.1.0
django-js-asset==1.1.0
django-pure-pagination==0.3.0
django-redis==4.9.0
django-reversion==3.0.2
django-taggit==0.23.0
et-xmlfile==1.0.1
future==0.15.2
gunicorn==19.9.0
httplib2==0.9.2
ipython==7.0.1
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.13.1
Markdown==3.0.1
#mysqlclient==1.3.13
odfpy==1.3.6
openpyxl==2.5.10
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.5
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.7
PyYAML==3.13
redis==2.10.6
simplegeneric==0.8.1
six==1.10.0
tablib==0.12.1
traitlets==4.3.2
unicodecsv==0.14.1
wcwidth==0.1.7
xlrd==1.1.0
xlwt==1.3.0
